requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
rembg==2.0.50
opencv-python==4.8.1.78
numpy==1.24.3
PyPDF2==3.0.1
reportlab==4.0.7
cryptography==41.0.7
python-dotenv==1.0.0
svgwrite==1.4.3
cairosvg==2.7.1
Werkzeug==3.0.1
